+++ library/src/schemas/record/record.ts
@@ -5,7 +5,7 @@ import type {
   ErrorMessage,
   InferIssue,
 } from '../../types/index.ts';
-import { _addIssue, _isAllowedObjectKey } from '../../utils/index.ts';
+import { _addIssue, _isValidObjectKey } from '../../utils/index.ts';
 import type {
   InferRecordInput,
   InferRecordOutput,
@@ -125,7 +125,7 @@ export function record(
           // Exclude blocked keys to prevent prototype pollutions
           // TODO: We should document that we exclude specific keys for
           // security reasons.
-          if (_isAllowedObjectKey(entryKey)) {
+          if (_isValidObjectKey(input, entryKey)) {
             // Get value of record entry
             const entryValue: unknown = input[entryKey as keyof typeof input];