+++ library/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Rename `_isAllowedObjectKey` to `_isValidObjectKey` and add check for inherited properties
+
 ## v0.32.0 (June 14, 2024)
 
 - Add `rawCheck`, `rawCheckAsync`, `rawTransform` and `rawTransformAsync` action (issue #597)